setup/pyproject.toml
[project]
name = "setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
]
